Index

MaxEntropyGraphs

MaxEntropyGraphs.AbstractMaxEntropyModel

MaxEntropyGraphs.ConvergenceError

Graphs.degree

MaxEntropyGraphs.ANND

MaxEntropyGraphs.ANND_in

MaxEntropyGraphs.ANND_out

MaxEntropyGraphs.M1

MaxEntropyGraphs.M10

MaxEntropyGraphs.M11

MaxEntropyGraphs.M12

MaxEntropyGraphs.M13

MaxEntropyGraphs.M2

MaxEntropyGraphs.M3

MaxEntropyGraphs.M4

MaxEntropyGraphs.M5

MaxEntropyGraphs.M6

MaxEntropyGraphs.M7

MaxEntropyGraphs.M8

MaxEntropyGraphs.M9

MaxEntropyGraphs.V_PB_parameters

MaxEntropyGraphs.V_motifs

MaxEntropyGraphs.V_p

MaxEntropyGraphs.a̲

MaxEntropyGraphs.a⭠

MaxEntropyGraphs.a⭢

MaxEntropyGraphs.a⭤

MaxEntropyGraphs.chesapeakebay

MaxEntropyGraphs.corporateclub

MaxEntropyGraphs.everglades

MaxEntropyGraphs.florida

MaxEntropyGraphs.indegree

MaxEntropyGraphs.instrength

MaxEntropyGraphs.littlerock

MaxEntropyGraphs.log_nan

MaxEntropyGraphs.maspalomas

MaxEntropyGraphs.np_unique_clone

MaxEntropyGraphs.outdegree

MaxEntropyGraphs.outstrength

MaxEntropyGraphs.parse_konect

MaxEntropyGraphs.readpajek

MaxEntropyGraphs.rhesus_macaques

MaxEntropyGraphs.squares

MaxEntropyGraphs.stmarks

MaxEntropyGraphs.strength

MaxEntropyGraphs.taro_exchange

MaxEntropyGraphs.triangles

MaxEntropyGraphs.wedges

Global

MaxEntropyGraphs — Module

MaxEntropyGraphs

Julia module for working with maximum entropy graphs

MaxEntropyGraphs.AbstractMaxEntropyModel — Type

AbstractMaxEntropyModel

An abstract type for a MaxEntropyModel. Each model has one or more structural constraints that are fixed while the rest of the network is completely random.

MaxEntropyGraphs.ConvergenceError — Type

ConvergenceError

Exception thrown when the optimisation method does not converge.

When using and optimisation method from the Optimisation.jl framework, the return code of the optimisation method is stored in the retcode field. When using the fixed point iteration method, the retcode field is set to nothing.

Utility fuctions

Some specific utility functions are made available:

MaxEntropyGraphs.log_nan — Function

log_nan(x::T)

Same as log(x), but returns NaN if x <= 0. Inspired by NaNMath.jl and https://github.com/JuliaMath/NaNMath.jl/issues/63. This methods is prefered over the ones from NaNMath.jl version because it does not require a foreign call expression to be evaluated, hence autodiff methods can be used with this.

Examples

julia> MaxEntropyGraphs.log_nan(10.)
2.302585092994046

julia> MaxEntropyGraphs.log_nan(-10.)
NaN

MaxEntropyGraphs.np_unique_clone — Function

np_unique_clone(x::Vector; sorted::Bool=false)

Julia replication of the numpy.unique(a, returncounts=True, returnindex=True, return_inverse=True) function from Python.

Returns a tuple of:

vector of unique values in x

vector of indices of the first occurence of each unique value in x. Follows the same order as the unique values.

vector of inverse indices of the original data in the unique values

vector of the counts of the unique values in x. Follows the same order as the unique values.

If sorted is true, the unique values are sorted by size and the other vectors are sorted accordingly.

Examples

julia> x = [1;2;2;4;1];

julia> np_unique_clone(x)
([1, 2, 4], [1, 2, 4], [1, 2, 2, 3, 1], [2, 2, 1])

julia> x = [10;9;9;8];

julia> np_unique_clone(x, sorted=true)
([8, 9, 10], [4, 2, 1], [3, 2, 2, 1], [1, 2, 1])

Small graph constructors

MaxEntropyGraphs.taro_exchange — Function

taro_exchange()

A small directed network that contains gift-givings (taro) between households in a Papuan village. A node represents a household and an edge between two households indicates that there happened a gift-giving. The network is directed and contains 22 nodes and 78 edges.

See also: KONECT

MaxEntropyGraphs.rhesus_macaques — Function

rhesus_maqaques()

A small weighted, directed network that contains observed grooming episodes between free ranging rhesus macaques (Macaca mulatta) in Cayo Santiago during a two month period in 1963. Cayo Santiago is an island off the coast of Puerto Rico, also known as Isla de los monos (Island of the monkeys). A node represents a monkey and a directed edge A → B denotes that the rhesus macaque A groomed rhesus macaque B. The integer edge weights indicate how often this behaviour was observed.

See also: KONECT

MaxEntropyGraphs.chesapeakebay — Function

chesapeakebay()

A small foodweb (directed network) of the Chesapeake Bay.

Original data available here

See also: MaxEntropyGraphs.readpajek

MaxEntropyGraphs.everglades — Function

everglades()

A small foodweb (directed network) of the Everglades.

Original data available here

See also: MaxEntropyGraphs.readpajek

MaxEntropyGraphs.florida — Function

florida()

A small foodweb (directed network) of the Florida Bay.

Original data available here

See also: MaxEntropyGraphs.readpajek

MaxEntropyGraphs.littlerock — Function

littlerock()

A small foodweb (directed network) of the Little Rock Lake.

Original data available here

See also: MaxEntropyGraphs.parse_konect

MaxEntropyGraphs.maspalomas — Function

maspalomas()

A small foodweb (directed network) from Charca de Maspalomas.

Original data available here

See also: MaxEntropyGraphs.readpajek

MaxEntropyGraphs.stmarks — Function

stmarks()

A small foodweb (directed network) from the St. Marks River.

Original data available here

See also: MaxEntropyGraphs.readpajek

MaxEntropyGraphs.parse_konect — Function

parse_konect(content::String)

Helper function to parse network data from the KONECT Project and return it to a graph object in the JuliaGraphs ecosystem.

Examples

# read the data from a file
G = open(io -> parse_konect(read(io, String)), "/path/to/KONECT/network/data")
# store the graph in Graphs.jl format
savegraph("path/to/my_KONECT_graph.lgz", G)

MaxEntropyGraphs.readpajek — Function

readpajek(f::String; is_directed::Bool=true)

helper function to load a network from the pajek dataset and return it to a graph object in the JuliaGraphs ecosystem.

For this simple helper, weights are discarded.

MaxEntropyGraphs.corporateclub — Function

corporateclub()

A bipartite network that contains membership information of corporate executive officers in social organizations such as clubs and boards. Nodes in the ⊥-layer represent persons and nodes in the ⊤-layer represent social organisations. An edge between a person and a social organization shows that the person is a member.

Original data available here

See also: MaxEntropyGraphs.parse_konect

Graph metrics

Graphs.degree — Function

degree(m::UBCM, i::Int; method=:reduced)

Return the expected degree for node i of the UBCM model m. Uses the reduced model parameters xᵣ for perfomance reasons.

Arguments

m::UBCM: the UBCM model

i::Int: the node for which to compute the degree.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = UBCM(MaxEntropyGraphs.Graphs.SimpleGraphs.smallgraph(:karate));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof([degree(model, 1), degree(model, 1, method=:full), degree(model, 1, method=:adjacency)])
Vector{Float64} (alias for Array{Float64, 1})

degree(m::UBCM[, v]; method=:reduced)

Return a vector corresponding to the expected degree of the UBCM model m each node. If v is specified, only return degrees for nodes in v.

Arguments

m::UBCM: the UBCM model

v::Vector{Int}: the nodes for which to compute the degree. Default is all nodes.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = UBCM(MaxEntropyGraphs.Graphs.SimpleGraphs.smallgraph(:karate));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof(degree(model, method=:adjacency)) 
Vector{Float64} (alias for Array{Float64, 1})

degree(m::DBCM, i::Int; method=:reduced)

In alignment with Graphs.jl, returns the sum of the expected out- and indegree for node i of the DBCM model m. Uses the reduced model parameters xᵣ for perfomance reasons.

Arguments

m::DBCM: the DBCM model

i::Int: the node for which to compute the degree.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = DBCM(MaxEntropyGraphs.Graphs.SimpleDiGraph(rhesus_macaques()));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof([degree(model, 1), degree(model, 1, method=:full), degree(model, 1, method=:adjacency)])
Vector{Float64} (alias for Array{Float64, 1})

degree(m::DBCM[, v]; method=:reduced)

In alignment with Graphs.jl, returns a vector corresponding to the sum of the expected out- and indegree of the DBCM model m each node. If v is specified, only return indegrees for nodes in v.

Arguments

m::DBCM: the DBCM model

v::Vector{Int}: the nodes for which to compute the degree. Default is all nodes.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = DBCM(MaxEntropyGraphs.Graphs.SimpleDiGraph(rhesus_macaques()));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof(degree(model, method=:adjacency)) 
Vector{Float64} (alias for Array{Float64, 1})

degree(m::BiCM, i::Int; method=:reduced)

Return the expected degree for node i of the BiCM model m. Uses the reduced model parameters xᵣ for perfomance reasons.

Arguments

m::BiCM: the BiCM model

i::Int: the node for which to compute the degree. This can be any of the nodes in the original graph used to define the model.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ, yᵣ, f⊥ and f⊤ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = BiCM(corporateclub());

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof([degree(model, 1), degree(model, 1, method=:full), degree(model, 1, method=:adjacency)])
Vector{Float64} (alias for Array{Float64, 1})

degree(m::BiCM[, v]; method=:reduced)

Return a vector corresponding to the expected degree of the BiCM model m each node. If v is specified, only return degrees for nodes in v.

Arguments

m::BiCM: the BiCM model

v::Vector{Int}: the nodes for which to compute the degree. Default is all nodes.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = BiCM(corporateclub());

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof(degree(model, method=:adjacency)) 
Vector{Float64} (alias for Array{Float64, 1})

MaxEntropyGraphs.outdegree — Function

outdegree(m::DBCM, i::Int; method=:reduced)

Return the expected degree vector for node i of the DBCM model m. Uses the reduced model parameters xᵣ for perfomance reasons.

Arguments

m::DBCM: the DBCM model

i::Int: the node for which to compute the degree.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = DBCM(MaxEntropyGraphs.Graphs.SimpleDiGraph(rhesus_macaques()));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof([outdegree(model, 1), outdegree(model, 1, method=:full), outdegree(model, 1, method=:adjacency)])
Vector{Float64} (alias for Array{Float64, 1})

outdegree(m::DBCM[, v]; method=:reduced)

Return a vector corresponding to the expected outdegree of the DBCM model m each node. If v is specified, only return outdegrees for nodes in v.

Arguments

m::DBCM: the DBCM model

v::Vector{Int}: the nodes for which to compute the outdegree. Default is all nodes.

method::Symbol: the method to use for computing the outdegree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = DBCM(MaxEntropyGraphs.Graphs.SimpleDiGraph(rhesus_macaques()));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof(outdegree(model, method=:adjacency)) 
Vector{Float64} (alias for Array{Float64, 1})

MaxEntropyGraphs.indegree — Function

indegree(m::DBCM, i::Int; method=:reduced)

Return the expected degree vector for node i of the DBCM model m. Uses the reduced model parameters xᵣ for perfomance reasons.

Arguments

m::DBCM: the DBCM model

i::Int: the node for which to compute the degree.

method::Symbol: the method to use for computing the degree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = DBCM(MaxEntropyGraphs.Graphs.SimpleDiGraph(rhesus_macaques()));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof([indegree(model, 1), indegree(model, 1, method=:full), indegree(model, 1, method=:adjacency)])
Vector{Float64} (alias for Array{Float64, 1})

indegree(m::DBCM[, v]; method=:reduced)

Return a vector corresponding to the expected indegree of the DBCM model m each node. If v is specified, only return indegrees for nodes in v.

Arguments

m::DBCM: the DBCM model

v::Vector{Int}: the nodes for which to compute the indegree. Default is all nodes.

method::Symbol: the method to use for computing the indegree. Can be any of the following:

:reduced (default) uses the reduced model parameters xᵣ for perfomance reasons.

:full uses all elements of the expected adjacency matrix.

:adjacency uses the precomputed adjacency matrix m.Ĝ of the model.

Examples

julia> model = DBCM(MaxEntropyGraphs.Graphs.SimpleDiGraph(rhesus_macaques()));

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> typeof(indegree(model, method=:adjacency)) 
Vector{Float64} (alias for Array{Float64, 1})

MaxEntropyGraphs.strength — Function

strength(G, T; dir)

Construct the strength vector for the graph G, filled with element type T and considering edge direction dir ∈ [:in, :out, :both] (default is :out).

strength(G, i, T; dir)

Construct the strength of node i for the graph G, filled with element type T and considering edge direction dir ∈ [:in, :out, :both] (default is :out).

MaxEntropyGraphs.outstrength — Function

instrength(G, T; dir)

Construct the outstrength vector for the graph G, filled with element type T.

MaxEntropyGraphs.instrength — Function

instrength(G, T; dir)

Construct the instrength vector for the graph G, filled with element type T.

MaxEntropyGraphs.ANND — Function

ANND(G::T, i::Int; check_directed::Bool=true) where {T<:Graphs.AbstractGraph}

Compute the average nearest neighbor degree (ANND) for node i in graph G. The ANND for a node i is defined as $ANND_i(A^{*}) = \frac{\sum_{j=1}^{N} a_{ij}k_j }{k_i}$ where $a_{ij}$ denotes the element of the adjacency matrix $A$ at row $i$ and column $j$, and $k_i$ denotes the degree of node $i$.

Notes:

the ANND is only defined for nodes with nonzero degree. If degree(G,i) = 0, then ANND(G,i) = 0.

if G is a directed graph, by default an error is thrown because the degree function returns the incoming plus outgoing edges for node i in this case. This can be turned off by setting check_directed=false.

Examples

julia> using Graphs

julia> G = smallgraph(:karate);

julia> ANND(G,1)
4.3125

julia> add_vertex!(G);

julia> ANND(G, nv(G))
0.0

julia> Gd = SimpleDiGraph(G);

julia> ANND(Gd,1)
ERROR: ArgumentError: The graph is directed. The degree function returns the incoming plus outgoing edges for node `i`. Consider using ANND_in or ANND_out instead.
[...]

julia> ANND(Gd,1, check_directed=false)
4.3125

See also: ANND_in, ANND_out, Graphs.degree

ANND(G::T, vs=vertices(G); check_directed::Bool=true) where {T<:Graphs.AbstractGraph}

Return a vector correcponding to the average nearest neighbor degree (ANND) all nodes in the graph G. If v is specified, only return the ANND for nodes in v. The ANND for a node i is defined as $ANND_i(A^{*}) = \frac{\sum_{j=1}^{N} a_{ij}k_j }{k_i}$ where $a_{ij}$ denotes the element of the adjacency matrix $A$ at row $i$ and column $j$, and $k_i$ denotes the degree of node $i$.

Notes:

the ANND is only defined for nodes with nonzero degree. If degree(G,i) = 0, then ANND(G,i) = 0.

if G is a directed graph, by default an error is thrown because the degree function returns the incoming plus outgoing edges for node i in this case.

This can be turned off by setting check_directed=false. This check is only performed once the actual computing.

Examples

julia> using Graphs

julia> G = smallgraph(:karate);

julia> ANND(G,[10; 20; 30])
3-element Vector{Float64}:
 13.5
 14.0
  9.0

julia> Gd = SimpleDiGraph(G);

julia> ANND(Gd,[10; 20; 30]);
ERROR: ArgumentError: The graph is directed. The degree function returns the incoming plus outgoing edges for node `i`. Consider using ANND_in or ANND_out instead.
[...]

julia> ANND(Gd,[10; 20; 30], check_directed=false)
3-element Vector{Float64}:
 13.5
 14.0
  9.0

See also: ANND_in, ANND_out, Graphs.degree

ANND(A::T, i::Int; check_dimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}

Compute the average nearest neighbor degree (ANND) for node i using adjacency matrix A. The ANND for a node i is defined as $ANND_i(A^{*}) = \frac{\sum_{j=1}^{N} a_{ij}k_j }{k_i}$ where $a_{ij}$ denotes the element of the adjacency matrix $A$ at row $i$ and column $j$, and $k_i$ denotes the degree of node $i$.

Notes:

this function is intented for use with the expected adjacency matrix of a ::AbstractMaxEntropyModel model. A separate method exists for ::AbstractGraph objects.

if A is not symmetrical, you have a directed graph, and this will throw an error by default. This can be turned off by setting check_directed=false.

the adjacency matrix should be square, if not, this will throw an error by default. This can be turned off by setting check_dimensions=false.

Examples

julia> using Graphs

julia> G = smallgraph(:karate);

julia> A = adjacency_matrix(G);

julia> ANND(A, 1)
4.3125

julia> Gd = SimpleDiGraph(G);

julia> add_vertex!(Gd); add_edge!(Gd, 1, nv(Gd));

julia> Ad = adjacency_matrix(Gd);

julia> ANND(Ad, 1)
ERROR: ArgumentError: The matrix is not symmetrical. Consider using ANND_in or ANND_out instead.
[...]

julia> ANND(Ad, 1, check_directed=false)
4.375

julia> ANND(rand(2,3),1)
ERROR: DimensionMismatch: `A` must be a square matrix.
[...]

See also: ANND_in, ANND_out, Graphs.degree

ANND(A::T, vs=1:size(A,1); check_dimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}

Return a vector correcponding to the average nearest neighbor degree (ANND) all nodes in the graph with adjacency matrix A. If v is specified, only return the ANND for nodes in v. The ANND for a node i is defined as $ANND_i(A^{*}) = \frac{\sum_{j=1}^{N} a_{ij}k_j }{k_i}$ where $a_{ij}$ denotes the element of the adjacency matrix $A$ at row $i$ and column $j$, and $k_i$ denotes the degree of node $i$.

Notes:

this function is intented for use with the expected adjacency matrix of a ::AbstractMaxEntropyModel model. A separate method exists for ::AbstractGraph objects.

if A is not symmetrical, you have a directed graph, and this will throw an error by default. This can be turned off by setting check_directed=false.

the adjacency matrix should be square, if not, this will throw an error by default. This can be turned off by setting check_dimensions=false.

Examples

julia> using Graphs

julia> G = smallgraph(:karate);

julia> A = adjacency_matrix(G);

julia> ANND(A);

julia> Gd = SimpleDiGraph(G);

julia> add_vertex!(Gd); add_edge!(Gd, 1, nv(Gd));

julia> Ad = adjacency_matrix(Gd);

julia> ANND(Ad)
ERROR: ArgumentError: The matrix is not symmetrical. Consider using ANND_in or ANND_out instead.
[...]

julia> ANND(Ad, check_directed=false)[1]
4.375

julia> ANND(rand(2,3))
ERROR: DimensionMismatch: `A` must be a square matrix.
[...]

See also: ANND_in, ANND_out, Graphs.degree

MaxEntropyGraphs.ANND_out — Function

ANNDout(A::T, vs=1:size(A,1); checkdimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}

Return a vector corresponding to the average nearest neighbor outdegree (ANND) all nodes in the graph with adjacency matrix A. If v is specified, only return the ANND_out for nodes in v.

See also: ANND_in, ANND

MaxEntropyGraphs.ANND_in — Function

ANNDin(A::T, vs=1:size(A,1); checkdimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}

Return a vector corresponding to the average nearest neighbor indegree (ANND) all nodes in the graph with adjacency matrix A. If v is specified, only return the ANND_in for nodes in v.

See also: ANND_out, ANND

MaxEntropyGraphs.wedges — Function

wedges(G::Graphs.SimpleGraph)
wedges(A::T; check_dimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}
wedges(m::UBCM)

Compute the number of (expected) wedges for an undirected graph. Can be done directly from the graph, based on the adjacency matrix or a UBCM model.

Arguments

For the adjacency matrix A, the following arguments can be passed:

check_dimensions: if true, check that A is a square matrix, otherwise throw an error.

check_directed: if true, check that A is symmetrical, otherwise throw an error.

Examples

julia> G = MaxEntropyGraphs.Graphs.smallgraph(:karate);

julia> model = UBCM(G);

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> (wedges(G), wedges(MaxEntropyGraphs.Graphs.adjacency_matrix(G)), wedges(model))
(528.0, 528.0, 528.0000011499742)

MaxEntropyGraphs.triangles — Function

triangles(G::Graphs.SimpleGraph)
triangles(A::T; check_dimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}
triangles(m::UBCM)

Compute the number of (expected) triangles for an undirected graph. Can be done directly from the graph, based on the adjacency matrix or a UBCM model.

Arguments

For the adjacency matrix A, the following arguments can be passed:

check_dimensions: if true, check that A is a square matrix, otherwise throw an error.

check_directed: if true, check that A is symmetrical, otherwise throw an error.

These checks can be turned off for perfomance reasons.

Examples

julia> G = MaxEntropyGraphs.Graphs.smallgraph(:karate);

julia> model = UBCM(G);

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> (triangles(G), triangles(MaxEntropyGraphs.Graphs.adjacency_matrix(G)), triangles(model))
(45, 45.0, 52.849301363026846)

MaxEntropyGraphs.squares — Function

squares(G::Graphs.SimpleGraph)
squares(A::T; check_dimensions::Bool=true, check_directed::Bool=true) where {T<:AbstractMatrix}
squares(m::UBCM)

Compute the number of (expected) squares for an undirected graph. Can be done directly from the graph, based on the adjacency matrix or a UBCM model.

Notes:

In this function, by $square$, a 'pure' square is understood, without any diagonals inside. This explains the difference with the induced subgraph count, which counts all squares, including those with triangles inside.

Arguments

For the adjacency matrix A, the following arguments can be passed:

check_dimensions: if true, check that A is a square matrix, otherwise throw an error.

check_directed: if true, check that A is symmetrical, otherwise throw an error.

These checks can be turned off for perfomance reasons.

Examples

julia> G = MaxEntropyGraphs.Graphs.smallgraph(:karate);

julia> model = UBCM(G);

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> (squares(G), squares(MaxEntropyGraphs.Graphs.adjacency_matrix(G)), squares(model))
(36.0, 36.0, 45.644736823949344)

MaxEntropyGraphs.a⭢ — Function

a⭢(A::T, i::Int, j::Int) where T<:AbstractArray

Compute non-recipocrated directed link from i to j and not from j to i.

MaxEntropyGraphs.a⭠ — Function

a⭠(A::T, i::Int, j::Int) where T<:AbstractArray

Computed non-recipocrated directed link not from i to j and from j to i.

MaxEntropyGraphs.a⭤ — Function

a⭤(A::T, i::Int, j::Int) where T<:AbstractArray

Computed recipocrated directed link between i and j.

MaxEntropyGraphs.a̲ — Function

a̲(A::T, i::Int, j::Int) where T<:AbstractArray

Compute absence of link between i and j.

MaxEntropyGraphs.M1 — Function

M1(A::T) where T<:AbstractArray

Count the occurence of motif M1 (Σ_{i≠j≠k} a⭠(i,j) a⭢(j,k) a̲(k,i) ) from the adjacency matrix.

M1(m::DBCM)

Count the occurence of motif M1 (Σ_{i≠j≠k} a⭠ (i,j) × a⭢ (j,k) × a̲ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M2 — Function

M2(A::T) where T<:AbstractArray

Count the occurence of motif M2 (Σ_{i≠j≠k} a⭠(i,j) a⭠(j,k) a̲(k,i) ) from the adjacency matrix.

M2(m::DBCM)

Count the occurence of motif M2 (Σ_{i≠j≠k} a⭠ (i,j) × a⭠ (j,k) × a̲ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M3 — Function

M3(A::T) where T<:AbstractArray

Count the occurence of motif M3 (Σ_{i≠j≠k} a⭠(i,j) a⭤(j,k) a̲(k,i) ) from the adjacency matrix.

M3(m::DBCM)

Count the occurence of motif M3 (Σ_{i≠j≠k} a⭠ (i,j) × a⭤ (j,k) × a̲ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M4 — Function

M4(A::T) where T<:AbstractArray

Count the occurence of motif M4 (Σ_{i≠j≠k} a⭠(i,j) a̲(j,k) a⭢(k,i) ) from the adjacency matrix.

M4(m::DBCM)

Count the occurence of motif M4 (Σ_{i≠j≠k} a⭠ (i,j) × a̲ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M5 — Function

M5(A::T) where T<:AbstractArray

Count the occurence of motif M5 (Σ_{i≠j≠k} a⭠(i,j) a⭢(j,k) a⭢(k,i) ) from the adjacency matrix.

M5(m::DBCM)

Count the occurence of motif M5 (Σ_{i≠j≠k} a⭠ (i,j) × a⭢ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M6 — Function

M6(A::T) where T<:AbstractArray

Count the occurence of motif M6 (Σ_{i≠j≠k} a⭠(i,j) a⭤(j,k) a⭢(k,i) ) from the adjacency matrix.

M6(m::DBCM)

Count the occurence of motif M6 (Σ_{i≠j≠k} a⭠ (i,j) × a⭤ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M7 — Function

M7(A::T) where T<:AbstractArray

Count the occurence of motif M7 (Σ_{i≠j≠k} a⭢(i,j) a⭤(j,k) a̲(k,i) ) from the adjacency matrix.

M7(m::DBCM)

Count the occurence of motif M7 (Σ_{i≠j≠k} a⭢ (i,j) × a⭤ (j,k) × a̲ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M8 — Function

M8(A::T) where T<:AbstractArray

Count the occurence of motif M8 (Σ_{i≠j≠k} a⭤(i,j) a⭤(j,k) a̲(k,i) ) from the adjacency matrix.

M8(m::DBCM)

Count the occurence of motif M8 (Σ_{i≠j≠k} a⭤ (i,j) × a⭤ (j,k) × a̲ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M9 — Function

M9(A::T) where T<:AbstractArray

Count the occurence of motif M9 (Σ_{i≠j≠k} a⭢(i,j) a⭢(j,k) a⭢(k,i) ) from the adjacency matrix.

M9(m::DBCM)

Count the occurence of motif M9 (Σ_{i≠j≠k} a⭢ (i,j) × a⭢ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M10 — Function

M10(A::T) where T<:AbstractArray

Count the occurence of motif M10 (Σ_{i≠j≠k} a⭤(i,j) a⭢(j,k) a⭢(k,i) ) from the adjacency matrix.

M10(m::DBCM)

Count the occurence of motif M10 (Σ_{i≠j≠k} a⭤ (i,j) × a⭢ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M11 — Function

M11(A::T) where T<:AbstractArray

Count the occurence of motif M11 (Σ_{i≠j≠k} a⭤(i,j) a⭠(j,k) a⭢(k,i) ) from the adjacency matrix.

M11(m::DBCM)

Count the occurence of motif M11 (Σ_{i≠j≠k} a⭤ (i,j) × a⭠ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M12 — Function

M12(A::T) where T<:AbstractArray

Count the occurence of motif M12 (Σ_{i≠j≠k} a⭤(i,j) a⭤(j,k) a⭢(k,i) ) from the adjacency matrix.

M12(m::DBCM)

Count the occurence of motif M12 (Σ_{i≠j≠k} a⭤ (i,j) × a⭤ (j,k) × a⭢ (k,i) ) from the DBCM model.

MaxEntropyGraphs.M13 — Function

M13(A::T) where T<:AbstractArray

Count the occurence of motif M13 (Σ_{i≠j≠k} a⭤(i,j) a⭤(j,k) a⭤(k,i) ) from the adjacency matrix.

M13(m::DBCM)

Count the occurence of motif M13 (Σ_{i≠j≠k} a⭤ (i,j) × a⭤ (j,k) × a⭤ (k,i) ) from the DBCM model.

MaxEntropyGraphs.V_motifs — Function

V_motifs(G::Graphs.SimpleGraph; membership::Vector=Graphs.bipartite_map(G), layer::Symbol=:bottom)

Count the total number of V-motif occurences in graph G for one of its layers.

Note: the bipartiteness of the graph is not explicitely checked.

Arguments

membership: the bipartite mapping of the graphs. This can be computed using Graphs.bipartite_map(G).

layer: the layer can be specified by passing layer=:bottom or layer=:top. Layer membership is determined by the bipartite map of the graph.

Examples

julia> using Graphs

julia> G = SimpleGraph(5); add_edge!(G, 1, 4); add_edge!(G, 2, 4); add_edge!(G, 3, 4); add_edge!(G, 3, 5);

julia> V_motifs(G, layer=:bottom)
3

julia> V_motifs(G, layer=:top)
1

V_motifs(A::T; layer::Symbol=:bottom, skipchecks::Bool=false) where {T<:AbstractMatrix}

Count the total number of V-motif occurence in the biadjacency matrix A for one of its layers.

Arguments

layer: the layer can be specified by passing layer=:bottom or layer=:top. Layer membership is determined by the bipartite map of the graph.

skipchecks: if true, skip the dimension check on A

Examples

julia> A = [1 1; 1 1; 1 0];

julia> V_motifs(A, layer=:bottom)
4

julia> V_motifs(A, layer=:top)
2

V_motifs(m::BiCM; layer::Symbol=:bottom)

Count the total number of V-motif occurences in the BiCM m for one of its layers.

Arguments

layer: the layer can be specified by passing layer=:bottom or layer=:top. Layer membership is determined by the bipartite map of the graph.

precomputed: if true, the expected values of the biadjacency matrix are used, otherwise the parameters are computed from the model parameters.

Examples

julia> model = BiCM(corporateclub());

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> V_motifs(model, layer=:bottom), V_motifs(model, layer=:bottom, precomputed=false)
(449.2569925909879, 449.2569925909879)

julia> V_motifs(model, layer=:top), V_motifs(model, layer=:top, precomputed=false)
(180.2569926636081, 180.2569926636081)

V_motifs(G::Graphs.SimpleGraph, i::Int, j::Int; membership::Vector=Graphs.bipartite_map(G))

Count the number of V-motif occurences in graph G between nodes i and j.

Notes:

the bipartiteness of the graph is not explicitely checked.

this uses the actual node numbers in the graph, not their numbering in the specific layer.

Arguments

membership: the bipartite mapping of the graphs. This can be computed using Graphs.bipartite_map(G).

Examples

julia> using Graphs

julia> G = SimpleGraph(5); add_edge!(G, 1, 4); add_edge!(G, 1, 5); add_edge!(G, 2, 4); add_edge!(G, 2, 5); add_edge!(G, 3, 4);

julia> V_motifs(G, 1, 2)
2

V_motifs(A::T, i::Int, j::Int; layer::Symbol=:bottom, skipchecks::Bool=false) where {T<:AbstractMatrix}

Count the number of V-motif occurences in the biadjacency matrix A between nodes i and j of a layer.

Arguments

layer: the layer can be specified by passing layer=:bottom or layer=:top.

skipchecks: if true, skip the dimension check on A

Notes: depending on the layer, the tuple (i, j) denotes rows (:bottom) or columns (:top) of the biadjacency matrix A.

Examples

julia> A = [1 1; 1 1; 1 0];

julia> V_motifs(A, 1, 2, layer=:bottom)
2

julia> V_motifs(A, 1, 2, layer=:top)
2

V_motifs(m::BiCM, i::Int, j::Int; layer::Symbol=:bottom, precomputed::Bool=true)

Count the number of expected V-motif occurences in the BiCM m between nodes i and j of a layer.

Arguments

layer: the layer can be specified by passing layer=:bottom or layer=:top.

precomputed: if true, the expected values of the biadjacency matrix are used, otherwise the parameters are computed from the model parameters.

Notes: depending on the layer, the tuple (i, j) denotes rows (:bottom) or columns (:top) of the expected biadjacency matrix A.

Examples

julia> model = BiCM(corporateclub());

julia> solve_model!(model);

julia> set_Ĝ!(model);

julia> V_motifs(model, 16, 13, layer=:bottom), V_motifs(model, 16, 13, layer=:bottom, precomputed=false)
(3.385652998856113, 3.385652998856112)

julia> V_motifs(model, 5, 1, layer=:top), V_motifs(model, 5, 1, layer=:top, precomputed=false)
(9.469880242964534, 9.469880242964534)

MaxEntropyGraphs.V_p — Function

V_p(m::BiCM, i::Int, j::Int; layer::Symbol=:bottom, precomputed::Bool=false)

Compute the p-value for the number of V-motifs between nodes i and j in the original graphs for the BiCM model m when projected onto the layer layer.

See also: V_motifs(::BiCM,::Int,::Int)

MaxEntropyGraphs.V_PB_parameters — Function

V_PB_parameters(m::BiCM, i::Int, j::Int; precomputed::Bool=false)

Compute the parameters of the Poisson-Binomial distribution for the number of V-motifs between nodes i and j for the BiCM model m.

Arguments

precomputed: if true, the expected values of the biadjacency matrix are used, otherwise the parameters are computed from the model parameters.

Notes:

layer membership is infered from node i

depending on the layer, the tuple (i, j) denotes rows (:bottom) or columns (:top) of the biadjacency matrix m.Ĝ.

Examples