Models

UBCM

MaxEntropyGraphs.AbstractMaxEntropyModel — Type

AbstractMaxEntropyModel

An abstract type for a MaxEntropyModel. Each model has one or more structural constraints that are fixed while the rest of the network is completely random.

MaxEntropyGraphs.DBCM — Type

DBCM{T,N} <: AbstractMaxEntropyModel where {T<:Real, N<:UInt}

Directed Binary Configuration Model (DBCM)

Maximum entropy model with a fixed in- and outdegree sequence.

MaxEntropyGraphs.UBCM — Type

UBCM{T,N} <: AbstractMaxEntropyModel where {T<:Real, N<:UInt}

Undirected Binary Configuration Model (UBCM)

Maximum entropy model with a fixed degree sequence. Uses the model where $x_i = e^{-\theta_i}$`

See also: AbstractMaxEntropyModel

Functions

MaxEntropyGraphs.∂UBCM_∂x! — Method

∂UBCM_∂x!(F::Vector, x::Vector, κ::Vector, f::Vector)

Gradient of likelihood function of the UBCM model using the x_i formulation.

F value of gradients x value of parameters κ value of (reduced degree vector) f frequency associated with each value in κ

see also UBCM`

MaxEntropyGraphs — Module

MaxEntropyGraphs

Julia module for working with maximum entropy graphs