Models

The different models that are available are described below. Each of these models is a subtype of an AbstractMaxEntropyModel

MaxEntropyGraphs.AbstractMaxEntropyModel — Type

AbstractMaxEntropyModel

An abstract type for a MaxEntropyModel. Each model has one or more structural constraints that are fixed while the rest of the network is completely random.

For each model there are multiple methods available to obtain the parameters that generate the maximum likelihood. Each of these models can be solved by using the solve! function.

solve!(model)

MaxEntropyGraphs.UBCM_∂ℒ_∂x! — Method

UBCM_∂ℒ_∂x!(F::Vector, x::Vector, κ::Vector, f::Vector)

Gradient of the likelihood function of the UBCM model using the $x_{i}$ formulation.

F value of gradients x value of parameters κ value of (reduced) degree vector f frequency associated with each value in κ

See also:

UBCM

MaxEntropyGraphs.UBCM_∂ℒ_∂x_it! — Method

UBCM_∂ℒ_∂x_it!(F::Vector, x::Vector, κ::Vector, f::Vector)

Iterative gradient of the likelihood function of the UBCM model using the x_i formulation.

F value of function x value of parameters κ value of (reduce) degree vector f frequency associated with each value in κ

See also:

UBCM

MaxEntropyGraphs.UBCM_∂ℒ_∂θ! — Method

UBCM_∂ℒ_∂x!(F::Vector, x::Vector, κ::Vector, f::Vector)

Gradient of the likelihood function of the UBCM model using the $x{i}$ formulation. Used to generate a function (F, x) -> UBCM_∂ℒ_∂x!(F, x, κ, f) by a UBCM instance.

F value of gradients x value of parameters κ value of (reduced degree vector) f frequency associated with each value in κ

see also:

UBCM

MaxEntropyGraphs.fffff — Method

fffff(x)

helper function for UBCM gradient computation

MaxEntropyGraphs.solve! — Method

solve!model::UBCM; , kwargs...)

Solve the equations of a UBCM model to obtain the maximum likelihood parameters.

See also UBCM

UBCM

Generate a Undirected Binary Configuration Model.

DBCM

Generate a Directed Binary Configuration Model.